Promote `CertificateType` enum to public API
Since 99447957 this has appeared in the public API, in the return types
of `server::ClientHello::server_cert_types()` and
`server::ClientHello::client_cert_types()`.

Move its definition into `crate::enums`.
