+++ rustls/src/msgs/enums.rs
@@ -313,19 +313,6 @@ enum_builder! {
 }
 
 enum_builder! {
-    /// The `CertificateType` enum sent in the cert_type extensions.
-    /// Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA.
-    ///
-    /// [RFC 6091 Section 5]: <https://datatracker.ietf.org/doc/html/rfc6091#section-5>
-    /// [RFC 7250 Section 7]: <https://datatracker.ietf.org/doc/html/rfc7250#section-7>
-    #[repr(u8)]
-    pub enum CertificateType {
-        X509 => 0x00,
-        RawPublicKey => 0x02,
-    }
-}
-
-enum_builder! {
     /// The Key Encapsulation Mechanism (`Kem`) type for HPKE operations.
     /// Listed by IANA, as specified in [RFC 9180 Section 7.1]
     ///
@@ -457,7 +444,6 @@ pub(crate) mod tests {
             CertificateStatusType::OCSP,
             CertificateStatusType::OCSP,
         );
-        test_enum8::<CertificateType>(CertificateType::X509, CertificateType::RawPublicKey);
     }
 
     pub(crate) fn test_enum8<T: for<'a> Codec<'a>>(first: T, last: T) {
